refactor: factor out common MockBackoffPolicy
