+++ google/cloud/internal/retry_loop_test.cc
@@ -14,6 +14,7 @@
 
 #include "google/cloud/internal/retry_loop.h"
 #include "google/cloud/options.h"
+#include "google/cloud/testing_util/mock_backoff_policy.h"
 #include "google/cloud/testing_util/status_matchers.h"
 #include <gmock/gmock.h>
 
@@ -23,6 +24,7 @@ GOOGLE_CLOUD_CPP_INLINE_NAMESPACE_BEGIN
 namespace internal {
 namespace {
 
+using ::google::cloud::testing_util::MockBackoffPolicy;
 using ::testing::ElementsAre;
 using ::testing::HasSubstr;
 using ::testing::Return;
@@ -96,12 +98,6 @@ TEST(RetryLoopTest, ReturnJustStatus) {
   OptionsSpan overlay(Options{}.set<StringOption>("uh-oh"));
   EXPECT_STATUS_OK(actual);
 }
-
-class MockBackoffPolicy : public BackoffPolicy {
- public:
-  MOCK_METHOD(std::unique_ptr<BackoffPolicy>, clone, (), (const, override));
-  MOCK_METHOD(std::chrono::milliseconds, OnCompletion, (), (override));
-};
 
 /**
  * @test Verify the backoff policy is queried after each failure.
