+++ google/cloud/pubsublite/CMakeLists.txt
@@ -234,7 +234,6 @@ function (google_cloud_cpp_pubsublite_client_define_tests)
         INTERFACE
             ${CMAKE_CURRENT_SOURCE_DIR}/testing/mock_alarm_registry.h
             ${CMAKE_CURRENT_SOURCE_DIR}/testing/mock_async_reader_writer.h
-            ${CMAKE_CURRENT_SOURCE_DIR}/testing/mock_backoff_policy.h
             ${CMAKE_CURRENT_SOURCE_DIR}/testing/mock_publisher.h
             ${CMAKE_CURRENT_SOURCE_DIR}/testing/mock_resumable_async_reader_writer_stream.h
             ${CMAKE_CURRENT_SOURCE_DIR}/testing/mock_retry_policy.h
